add haskell modules support.
* it fails a lot, I think it is because some module names overlap,
my fetching process should somehow account for this and store both
or do some other kind of magic (for example, prioritizing the 'base'
package).
* also, in my quick tests I've encountered at least one conflict:
the 'Config' module is probably a common name, and it is pointing to
some arbitrary package that wasn't listed in the stack.yaml file for
the package I was browsing.
* some modules are missing entirely also, it is probably because they
are present only in some old versions of hackage packages (I'm only
fetching the latest versions), or maybe there's something wrong with my
fetching process.