attempt to fix the back button problem.
if the cause is that GitHub's own pjax is interfering
with the normal page load event (and specially with
chrome.webNavigation.onHistoryStateUpdated events) then
this should fix it.

there should be no problem in running `main()` twice,
since it has its own deduplicating mechanism which costs
nothing. even if the deduplicating somehow doesn't work
(because GitHub modifies the DOM?) it still matter little,
because all XHR requests are cached.

closes https://github.com/fiatjaf/module-linker/issues/10